Add course detail with assignment cards
