+++ iOS/Views/Course/HorizontalCourseList.swift
@@ -7,18 +7,30 @@
 
 import SwiftUI
 
-struct CurrentCourseList: View {
+struct HorizontalCourseList: View {
     var courses: [Course]
     
     var body: some View {
-        NavigationView {
-            VerticalCourseList(courses: courses)
-            .navigationBarTitle(Text("Current Courses"))
+        ScrollView(.horizontal) {
+            LazyHStack {
+                ForEach(courses) { course in
+                    NavigationLink(destination: CourseDetail(course: course)){
+                        CourseCard(
+                            courseName: course.name,
+                            school: course.school,
+                            credits: course.credits,
+                            grade: course.grade,
+                            goal: course.goal
+                        )
+                        .padding([.vertical, .trailing])
+                    }
+                }
+            }
         }
     }
 }
 
-struct CurrentCourseList_Previews: PreviewProvider {
+struct HorizontalCourseList_Previews: PreviewProvider {
     private static let courses = [
         Course(
             id: 1,
@@ -26,7 +38,8 @@ struct CurrentCourseList_Previews: PreviewProvider {
             credits: 0.5,
             grade: Grade(weightAchieved: 45, weightLost: 10),
             goal: Grade(percentage: 90),
-            school: School(name: .UW)
+            school: School(name: .UW),
+            assignments: []
         ),
         Course(
             id: 2,
@@ -34,7 +47,8 @@ struct CurrentCourseList_Previews: PreviewProvider {
             credits: 0.5,
             grade: Grade(weightAchieved: 35, weightLost: 20),
             goal: Grade(percentage: 70),
-            school: School(name: .WLU)
+            school: School(name: .WLU),
+            assignments: []
         ),
         Course(
             id: 3,
@@ -42,7 +56,8 @@ struct CurrentCourseList_Previews: PreviewProvider {
             credits: 0.5,
             grade: Grade(weightAchieved: 90, weightLost: 10),
             goal: Grade(percentage: 50),
-            school: School(name: .UW)
+            school: School(name: .UW),
+            assignments: []
         ),
         Course(
             id: 4,
@@ -50,7 +65,8 @@ struct CurrentCourseList_Previews: PreviewProvider {
             credits: 1.0,
             grade: Grade(weightAchieved: 50, weightLost: 10),
             goal: Grade(percentage: 80),
-            school: School(name: .WLU)
+            school: School(name: .WLU),
+            assignments: []
         ),
         Course(
             id: 5,
@@ -58,10 +74,12 @@ struct CurrentCourseList_Previews: PreviewProvider {
             credits: 0.5,
             grade: Grade(weightAchieved: 30, weightLost: 70),
             goal: Grade(percentage: 10),
-            school: School(name: .WLU)
+            school: School(name: .WLU),
+            assignments: []
         )
     ]
     static var previews: some View {
-        CurrentCourseList(courses: courses)
+        HorizontalCourseList(courses: courses)
+        
     }
 }
