+++ iOS/Views/Course/VerticalCourseList.swift
@@ -7,12 +7,12 @@
 
 import SwiftUI
 
-struct HorizontalCourseList: View {
+struct VerticalCourseList: View {
     var courses: [Course]
     
     var body: some View {
-        ScrollView(.horizontal) {
-            LazyHStack {
+        ScrollView(.vertical) {
+            LazyVStack {
                 ForEach(courses) { course in
                     NavigationLink(destination: CourseDetail(course: course)){
                         CourseCard(
@@ -22,15 +22,15 @@ struct HorizontalCourseList: View {
                             grade: course.grade,
                             goal: course.goal
                         )
-                        .padding([.vertical, .trailing])
                     }
                 }
+                .padding([.horizontal, .bottom])
             }
         }
     }
 }
 
-struct HorizontalCourseList_Previews: PreviewProvider {
+struct VerticalCourseList_Previews: PreviewProvider {
     private static let courses = [
         Course(
             id: 1,
@@ -38,7 +38,8 @@ struct HorizontalCourseList_Previews: PreviewProvider {
             credits: 0.5,
             grade: Grade(weightAchieved: 45, weightLost: 10),
             goal: Grade(percentage: 90),
-            school: School(name: .UW)
+            school: School(name: .UW),
+            assignments: []
         ),
         Course(
             id: 2,
@@ -46,7 +47,8 @@ struct HorizontalCourseList_Previews: PreviewProvider {
             credits: 0.5,
             grade: Grade(weightAchieved: 35, weightLost: 20),
             goal: Grade(percentage: 70),
-            school: School(name: .WLU)
+            school: School(name: .WLU),
+            assignments: []
         ),
         Course(
             id: 3,
@@ -54,7 +56,8 @@ struct HorizontalCourseList_Previews: PreviewProvider {
             credits: 0.5,
             grade: Grade(weightAchieved: 90, weightLost: 10),
             goal: Grade(percentage: 50),
-            school: School(name: .UW)
+            school: School(name: .UW),
+            assignments: []
         ),
         Course(
             id: 4,
@@ -62,7 +65,8 @@ struct HorizontalCourseList_Previews: PreviewProvider {
             credits: 1.0,
             grade: Grade(weightAchieved: 50, weightLost: 10),
             goal: Grade(percentage: 80),
-            school: School(name: .WLU)
+            school: School(name: .WLU),
+            assignments: []
         ),
         Course(
             id: 5,
@@ -70,11 +74,11 @@ struct HorizontalCourseList_Previews: PreviewProvider {
             credits: 0.5,
             grade: Grade(weightAchieved: 30, weightLost: 70),
             goal: Grade(percentage: 10),
-            school: School(name: .WLU)
+            school: School(name: .WLU),
+            assignments: []
         )
     ]
     static var previews: some View {
-        HorizontalCourseList(courses: courses)
-        
+        VerticalCourseList(courses: courses)
     }
 }
