+++ iOS/Views/Home.swift
@@ -1,5 +1,5 @@
 //
-//  SemesterList.swift
+//  Home.swift
 //  Grade Tracker
 //
 //  Created by Matthew Froggatt on 2021-03-18.
@@ -7,25 +7,24 @@
 
 import SwiftUI
 
-struct SemesterList: View {
+struct Home: View {
     var semesters: [Semester]
-    
+    var courses: [Course]
     var body: some View {
-        NavigationView {
-            List(semesters) { semester in
-                VStack(alignment: .leading) {
-                    Text(semester.name)
-                        .font(.headline)
-                        .foregroundColor(.primary)
-                    HorizontalCourseList(courses: semester.courses)
+        TabView {
+            CurrentCourseList(courses: courses)
+                .tabItem {
+                    Label("Courses", systemImage: "star")
                 }
-            }
-            .navigationBarTitle("All Semesters")
+            SemesterList(semesters: semesters)
+                .tabItem {
+                    Label("Semesters", systemImage: "list.bullet")
+                }
         }
     }
 }
 
-struct SemesterList_Previews: PreviewProvider {
+struct Home_Previews: PreviewProvider {
     private static let semesters = [
         Semester(
             id: 1,
@@ -37,7 +36,16 @@ struct SemesterList_Previews: PreviewProvider {
                     credits: 0.5,
                     grade: Grade(weightAchieved: 45, weightLost: 10),
                     goal: Grade(percentage: 90),
-                    school: School(name: .UW)
+                    school: School(name: .UW),
+                    assignments: [
+                        Assignment(id: 1, name: "Test 1", weight: 15),
+                        Assignment(
+                            id: 2,
+                            name: "Test 2",
+                            weight: 30,
+                            grade: Grade(percentage: 15)
+                        )
+                    ]
                 ),
                 Course(
                     id: 2,
@@ -45,7 +53,8 @@ struct SemesterList_Previews: PreviewProvider {
                     credits: 0.5,
                     grade: Grade(weightAchieved: 35, weightLost: 20),
                     goal: Grade(percentage: 70),
-                    school: School(name: .WLU)
+                    school: School(name: .WLU),
+                    assignments: []
                 ),
                 Course(
                     id: 3,
@@ -53,7 +62,8 @@ struct SemesterList_Previews: PreviewProvider {
                     credits: 0.5,
                     grade: Grade(weightAchieved: 90, weightLost: 10),
                     goal: Grade(percentage: 50),
-                    school: School(name: .UW)
+                    school: School(name: .UW),
+                    assignments: []
                 ),
                 Course(
                     id: 4,
@@ -61,7 +71,8 @@ struct SemesterList_Previews: PreviewProvider {
                     credits: 1.0,
                     grade: Grade(weightAchieved: 50, weightLost: 10),
                     goal: Grade(percentage: 80),
-                    school: School(name: .WLU)
+                    school: School(name: .WLU),
+                    assignments: []
                 ),
                 Course(
                     id: 5,
@@ -69,13 +80,15 @@ struct SemesterList_Previews: PreviewProvider {
                     credits: 0.5,
                     grade: Grade(weightAchieved: 30, weightLost: 70),
                     goal: Grade(percentage: 10),
-                    school: School(name: .WLU)
+                    school: School(name: .WLU),
+                    assignments: []
                 )
             ]
         )
     ]
     
+    private static let courses = semesters[0].courses
     static var previews: some View {
-        SemesterList(semesters: semesters)
+        Home(semesters: semesters, courses: courses)
     }
 }
